backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.0
python-dotenv==1.2.1
gunicorn==22.0.0
requests==2.32.4
cryptography==46.0.5
PyJWT==2.12.0
urllib3==2.6.3
